Static IP form for BMC host discovery
