+++ libs/ui-lib/lib/common/components/index.ts
@@ -9,3 +9,4 @@ export * from './hosts';
 export * from './fetching';
 export * from './AddHosts';
 export * from './featureSupportLevels';
+export * from './staticIP';
