+++ libs/ui-lib/lib/common/components/staticIP/index.tsx
@@ -1,0 +1,6 @@
+export * from './MachineNetworkField';
+export * from './validationSchemas';
+export * from './types';
+export * from './IPConfigFields';
+export * from './machineNetwork';
+export * from './protocolVersion';
