+++ libs/ui-lib/lib/common/components/staticIP/machineNetwork.ts
@@ -1,4 +1,4 @@
-import { Cidr } from './dataTypes';
+import { Cidr } from './types';
 
 export const getMachineNetworkCidr = (machineNetwork: Cidr) => {
   return `${machineNetwork.ip}/${machineNetwork.prefixLength}`;
