+++ libs/ui-lib/lib/ocm/components/clusterConfiguration/staticIp/components/FormViewNetworkWide/FormViewNetworkWideFields.tsx
@@ -2,53 +2,36 @@ import React from 'react';
 import {
   Text,
   TextVariants,
-  Grid,
   FormGroup,
-  TextInputTypes,
   Alert,
   AlertVariant,
   TextContent,
-  Flex,
-  FlexItem,
   ButtonVariant,
-  FormHelperText,
-  HelperText,
-  HelperTextItem,
 } from '@patternfly/react-core';
 import { useField, useFormikContext } from 'formik';
 import {
   ConfirmationModal,
   getFieldId,
-  getHumanizedSubnetRange,
-  PopoverIcon,
-} from '../../../../../../common';
-import {
-  getAddressObject,
   getProtocolVersionLabel,
-  getShownProtocolVersions,
-} from '../../data/protocolVersion';
+  IpConfig,
+  IpConfigFields,
+  MIN_VLAN_ID,
+  MAX_VLAN_ID,
+  PopoverIcon,
+  useTranslation,
+} from '../../../../../../common';
+import { getShownProtocolVersions } from '../../../../../../common/components/staticIP/protocolVersion';
 import * as types from '../../data/dataTypes';
 import { getEmptyIpConfig } from '../../data/emptyData';
 import {
-  Cidr,
   FormViewHost,
   FormViewNetworkWideValues,
-  IpConfig,
   ProtocolVersion,
   StaticProtocolType,
 } from '../../data/dataTypes';
-import { getMachineNetworkCidr } from '../../data/machineNetwork';
-import useFieldErrorMsg from '../../../../../../common/hooks/useFieldErrorMsg';
-import {
-  MIN_PREFIX_LENGTH,
-  MAX_PREFIX_LENGTH,
-  MAX_VLAN_ID,
-  MIN_VLAN_ID,
-} from './formViewNetworkWideValidationSchema';
 import { OcmCheckboxField, OcmInputField, OcmRadioField } from '../../../../ui/OcmFormFields';
 
 import '../staticIp.css';
-import ExclamationCircleIcon from '@patternfly/react-icons/dist/js/icons/exclamation-circle-icon';
 
 const hostsConfiguredAlert = (
   <Alert
@@ -59,106 +42,13 @@ const hostsConfiguredAlert = (
   />
 );
 
-const MachineNetwork: React.FC<{ fieldName: string; protocolVersion: ProtocolVersion }> = ({
-  fieldName,
-  protocolVersion,
-}) => {
-  const [{ value }] = useField<Cidr>(fieldName);
-  const ipFieldName = `${fieldName}.ip`;
-  const prefixLengthFieldName = `${fieldName}.prefixLength`;
-  const ipErrorMessage = useFieldErrorMsg({ name: ipFieldName });
-  const prefixLengthErrorMessage = useFieldErrorMsg({ name: prefixLengthFieldName });
-  const errorMessage = ipErrorMessage || prefixLengthErrorMessage;
-  const machineNetworkHelptext = React.useMemo(() => {
-    if (errorMessage) {
-      return '';
-    }
-    const cidr = getMachineNetworkCidr(value);
-    return getHumanizedSubnetRange(getAddressObject(cidr, protocolVersion));
-  }, [value, protocolVersion, errorMessage]);
-  const fieldId = getFieldId(`${fieldName}`, 'input');
-  return (
-    <FormGroup
-      labelIcon={
-        <PopoverIcon noVerticalAlign bodyContent="The range of IP addresses of the hosts." />
-      }
-      label="Subnet"
-      fieldId={fieldId}
-      isRequired
-      className="subnet"
-    >
-      <Flex>
-        <FlexItem spacer={{ default: 'spacerSm' }}>
-          <OcmInputField
-            name={`${fieldName}.ip`}
-            isRequired={true}
-            data-testid={`${protocolVersion}-machine-network-ip`}
-            showErrorMessage={false}
-          />
-        </FlexItem>
-        <FlexItem spacer={{ default: 'spacerSm' }}>{'/'}</FlexItem>
-        <FlexItem>
-          <OcmInputField
-            name={`${fieldName}.prefixLength`}
-            isRequired={true}
-            data-testid={`${protocolVersion}-machine-network-prefix-length`}
-            type={TextInputTypes.number}
-            showErrorMessage={false}
-            min={MIN_PREFIX_LENGTH}
-            max={
-              protocolVersion === ProtocolVersion.ipv4
-                ? MAX_PREFIX_LENGTH.ipv4
-                : MAX_PREFIX_LENGTH.ipv6
-            }
-          />
-        </FlexItem>
-      </Flex>
-      {(errorMessage || machineNetworkHelptext) && (
-        <FormHelperText>
-          <HelperText>
-            <HelperTextItem
-              icon={errorMessage ? <ExclamationCircleIcon /> : null}
-              variant={errorMessage ? 'error' : 'default'}
-              id={errorMessage ? `${fieldId}-helper-error` : `${fieldId}-helper`}
-            >
-              {errorMessage ? errorMessage : machineNetworkHelptext}
-            </HelperTextItem>
-          </HelperText>
-        </FormHelperText>
-      )}
-    </FormGroup>
-  );
-};
-
-const IpConfigFields: React.FC<{
-  fieldName: string;
-  protocolVersion: ProtocolVersion;
-}> = ({ protocolVersion, fieldName }) => {
-  return (
-    <Grid hasGutter>
-      <MachineNetwork fieldName={`${fieldName}.machineNetwork`} protocolVersion={protocolVersion} />
-      <OcmInputField
-        isRequired
-        label="Default gateway"
-        labelIcon={
-          <PopoverIcon
-            noVerticalAlign
-            bodyContent="An IP address to where any IP packet should be forwarded in case there is no other routing rule configured for a destination IP."
-          />
-        }
-        name={`${fieldName}.gateway`}
-        data-testid={`${protocolVersion}-gateway`}
-      />
-    </Grid>
-  );
-};
-
 const ipv6ValuesEmpty = (values: FormViewNetworkWideValues) =>
   values.ipConfigs.ipv6.gateway === '' &&
   values.ipConfigs.ipv6.machineNetwork.ip === '' &&
   values.ipConfigs.ipv6.machineNetwork.prefixLength === '';
 
 export const ProtocolTypeSelect = () => {
+  const { t } = useTranslation();
   const selectFieldName = 'protocolType';
   const [{ value: protocolType }, , { setValue: setProtocolType }] =
     useField<StaticProtocolType>(selectFieldName);
@@ -192,7 +82,7 @@ export const ProtocolTypeSelect = () => {
         <OcmRadioField
           label={
             <>
-              {getProtocolVersionLabel(ProtocolVersion.ipv4)}{' '}
+              {getProtocolVersionLabel(ProtocolVersion.ipv4, t)}{' '}
               <PopoverIcon
                 noVerticalAlign
                 bodyContent="Select this when your hosts are using only IPv4."
@@ -301,16 +191,11 @@ export const FormViewNetworkWideFields = ({ hosts }: { hosts: FormViewHost[] }) 
       />
 
       {getShownProtocolVersions(values.protocolType).map((protocolVersion) => (
-        <FormGroup
+        <IpConfigFields
+          fieldName={`ipConfigs.${protocolVersion}`}
+          protocolVersion={protocolVersion}
           key={protocolVersion}
-          label={getProtocolVersionLabel(protocolVersion)}
-          fieldId={getFieldId(`ip-configs-${protocolVersion}`, 'input')}
-        >
-          <IpConfigFields
-            fieldName={`ipConfigs.${protocolVersion}`}
-            protocolVersion={protocolVersion}
-          />
-        </FormGroup>
+        />
       ))}
     </>
   );
